% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assistant_create.R
\name{assistant_create}
\alias{assistant_create}
\title{Create an Assistant}
\usage{
assistant_create(instructions, name, model, api_key)
}
\arguments{
\item{instructions}{The instructions for the assistant}

\item{name}{The name of the assistant}

\item{model}{The model version (e.g., "gpt-4o")}

\item{api_key}{The OpenAI Api key}
}
\value{
An Assistant object. See https://platform.openai.com/docs/api-reference/assistants/object
}
\description{
Create an Assistant
}
\examples{
assistant_create()
}
