% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assistant_get.R
\name{assistant_get}
\alias{assistant_get}
\title{Retrieve an assistant}
\usage{
assistant_get(assistant_id, token)
}
\arguments{
\item{assistant_id}{The ChatGPT assistant ID}

\item{token}{A ChatGPT token}
}
\value{
A list containing the assistants parameters
}
\description{
Retrieve an assistant
}
\examples{
See ChatGPT
}
