% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assistant_list.R
\name{assistant_list}
\alias{assistant_list}
\title{List assistants}
\usage{
assistant_list(token)
}
\arguments{
\item{token}{A ChatGPT token}
}
\value{
A list with all the available ChatGPT assistants
}
\description{
List assistants
}
\examples{
See ChatGPT
}
