% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/message_create.R
\name{message_create}
\alias{message_create}
\title{Create a new message}
\usage{
message_create(thread_id, message, token)
}
\arguments{
\item{thread_id}{The thread id}

\item{message}{The new message}

\item{token}{The token}
}
\value{
A message list
}
\description{
Create a new message
}
\examples{
See ChatGPT
}
