% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/message_delete.R
\name{message_delete}
\alias{message_delete}
\title{Delete a message}
\usage{
message_delete(thread_id, message_id, token)
}
\arguments{
\item{thread_id}{The thread ID}

\item{message_id}{The message ID}

\item{token}{The ChatGPT Token}
}
\value{
Object
}
\description{
Delete a message
}
\examples{
See ChatGPT
}
