% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/message_get.R
\name{message_get}
\alias{message_get}
\title{Get a specific message}
\usage{
message_get(thread_id, message_id, token)
}
\arguments{
\item{thread_id}{The thread id}

\item{message_id}{The message id}

\item{token}{The token}
}
\value{
Object
}
\description{
Get a specific message
}
\examples{
See ChatGPT
}
