% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/message_list.R
\name{message_list}
\alias{message_list}
\title{Get messages}
\usage{
message_list(thread_id, token)
}
\arguments{
\item{thread_id}{Thread id}

\item{token}{A Token}
}
\value{
Object
}
\description{
Get messages
}
\examples{
See ChatGPT
}
