% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/message_manage.R
\name{message_manage}
\alias{message_manage}
\title{Manage Messages}
\usage{
message_manage(thread_id, api_key)
}
\arguments{
\item{thread_id}{The Thread ID}

\item{api_key}{The OpenAI Api key}
}
\value{
Print the status
}
\description{
Manage Messages
}
\examples{
See platform.openai.com/docs/api-reference/messages
}
