% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_create.R
\name{run_create}
\alias{run_create}
\title{Create run}
\usage{
run_create(thread_id, asst_id, token)
}
\arguments{
\item{thread_id}{thread ID}

\item{asst_id}{assistant id}

\item{token}{token}
}
\value{
A run
}
\description{
Create run
}
\examples{
See chatgpt
}
