% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_get.R
\name{run_get}
\alias{run_get}
\title{Get a run}
\usage{
run_get(thread_id, run_id, token)
}
\arguments{
\item{thread_id}{Thread id}

\item{run_id}{Run id}

\item{token}{Token}
}
\value{
A run
}
\description{
Get a run
}
\examples{
See chatgpt
}
