% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_go.R
\name{run_go}
\alias{run_go}
\title{Execture run}
\usage{
run_go(thread_id, run_id, token)
}
\arguments{
\item{thread_id}{Thread id}

\item{run_id}{Run id}

\item{token}{Token}
}
\value{
A run
}
\description{
Execture run
}
\examples{
See chatgpt
}
