% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_list.R
\name{run_list}
\alias{run_list}
\title{List run status}
\usage{
run_list(thread_id, token)
}
\arguments{
\item{thread_id}{Thread Id}

\item{token}{Token}
}
\value{
A run
}
\description{
List run status
}
\examples{
See chatgpt
}
