% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/thread_delete.R
\name{thread_delete}
\alias{thread_delete}
\title{Delete a thread}
\usage{
thread_delete(thread_id, token)
}
\arguments{
\item{thread_id}{The thread ID}

\item{token}{The token}
}
\value{
Pink
}
\description{
Delete a thread
}
\examples{
See ChatGPT
}
