% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/thread_get.R
\name{thread_get}
\alias{thread_get}
\title{Get a thread}
\usage{
thread_get(thread_id, token)
}
\arguments{
\item{thread_id}{The ID thread}
}
\value{
A thread object
}
\description{
Get a thread
}
\examples{
See ChatGPT
}
