% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/thread_manage.R
\name{thread_manage}
\alias{thread_manage}
\title{Create and store a new therad}
\usage{
thread_manage(thread_path, api_key)
}
\arguments{
\item{thread_path}{The path to the thread object}

\item{api_key}{The OpenAI Api key}
}
\value{
An Thread object. See https://platform.openai.com/docs/api-reference/assistants/thread
}
\description{
Create and store a new therad
}
