% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assistant_get.R
\name{assistant_get}
\alias{assistant_get}
\title{Retrieve an Assistant}
\usage{
assistant_get(assistant_id, api_key)
}
\arguments{
\item{assistant_id}{The Assistant ID}

\item{api_key}{The OpenAI Api key}
}
\value{
An Assistant object. See https://platform.openai.com/docs/api-reference/assistants/object
}
\description{
Retrieve an Assistant
}
