% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assistant_list.R
\name{assistant_list}
\alias{assistant_list}
\title{List Assistants}
\usage{
assistant_list(api_key)
}
\arguments{
\item{api_key}{The OpenAI Api key}
}
\value{
A list of Assistants objects. See https://platform.openai.com/docs/api-reference/assistants/object
}
\description{
List Assistants
}
\examples{
assistant_list()
}
