% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/message_create.R
\name{message_create}
\alias{message_create}
\title{Create a Message}
\usage{
message_create(thread_id, message, api_key)
}
\arguments{
\item{thread_id}{The Thread ID}

\item{message}{The message}

\item{api_key}{The OpenAI Api key}
}
\value{
A Message object. See https://platform.openai.com/docs/api-reference/messages/object
}
\description{
Create a Message
}
\examples{
See https://platform.openai.com/docs/api-reference/messages/createMessage
}
