% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/message_get.R
\name{message_get}
\alias{message_get}
\title{Retrieve a Message}
\usage{
message_get(thread_id, message_id, api_key)
}
\arguments{
\item{thread_id}{The Thread ID}

\item{message_id}{The Message ID}

\item{api_key}{The OpenAI Api key}
}
\value{
A Message object. See https://platform.openai.com/docs/api-reference/messages/object
}
\description{
Retrieve a Message
}
\examples{
See https://platform.openai.com/docs/api-reference/messages/getMessage
}
