% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/message_list.R
\name{message_list}
\alias{message_list}
\title{List Messages}
\usage{
message_list(thread_id, api_key)
}
\arguments{
\item{thread_id}{The Thread ID}

\item{api_key}{The OpenAI Api key}
}
\value{
A list of Message objects. See https://platform.openai.com/docs/api-reference/messages/object
}
\description{
List Messages
}
