% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_create.R
\name{run_create}
\alias{run_create}
\title{Create Run}
\usage{
run_create(thread_id, assistant_id, api_key)
}
\arguments{
\item{thread_id}{The Thread ID}

\item{assistant_id}{The Assistant ID}

\item{api_key}{The OpenAI Api key}
}
\value{
A Run object. See https://platform.openai.com/docs/api-reference/runs/object
}
\description{
Create Run
}
\examples{
See https://platform.openai.com/docs/api-reference/runs/createRun
}
