% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_get.R
\name{run_get}
\alias{run_get}
\title{Retrieve a Run}
\usage{
run_get(thread_id, run_id, api_key)
}
\arguments{
\item{thread_id}{The Thread ID}

\item{run_id}{The Run ID}

\item{api_key}{The OpenAI Api key}
}
\value{
A Run object. See https://platform.openai.com/docs/api-reference/runs/object
}
\description{
Retrieve a Run
}
\examples{
See platform.openai.com/docs/api-reference/runs/getRun
}
