% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_go.R
\name{run_go}
\alias{run_go}
\title{Execture a Run}
\usage{
run_go(thread_id, run_id, api_key)
}
\arguments{
\item{thread_id}{The Thread ID}

\item{run_id}{The Run ID}

\item{api_key}{The OpenAI Api key}
}
\value{
A Run object. See https://platform.openai.com/docs/api-reference/runs/object
}
\description{
Execture a Run
}
\examples{
run_go()
}
