% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_list.R
\name{run_list}
\alias{run_list}
\title{List Runs}
\usage{
run_list(thread_id, api_key)
}
\arguments{
\item{thread_id}{The Thread ID}

\item{api_key}{The OpenAI Api key}
}
\value{
A list of Run objects. See https://platform.openai.com/docs/api-reference/runs/object
}
\description{
List Runs
}
\examples{
See https://platform.openai.com/docs/api-reference/runs/listRuns
}
