% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/thread_create.R
\name{thread_create}
\alias{thread_create}
\title{Create a Thread}
\usage{
thread_create(api_key)
}
\arguments{
\item{api_key}{The OpenAI Api key}
}
\value{
A Thread object. See https://platform.openai.com/docs/api-reference/threads/object
}
\description{
Create a Thread
}
\examples{
thread_create()
}
