% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/thread_delete.R
\name{thread_delete}
\alias{thread_delete}
\title{Delete a Thread}
\usage{
thread_delete(thread_id, api_key)
}
\arguments{
\item{thread_id}{The Thread ID}

\item{api_key}{The OpenAI Api key}
}
\value{
Information on the deleted Thread object. See https://platform.openai.com/docs/api-reference/threads/object
}
\description{
Delete a Thread
}
\examples{
See https://platform.openai.com/docs/api-reference/threads/deleteThread
}
