% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/thread_get.R
\name{thread_get}
\alias{thread_get}
\title{Retrieve a Thread}
\usage{
thread_get(thread_id, api_key)
}
\arguments{
\item{thread_id}{The Thread ID}

\item{api_key}{The OpenAI Api key}
}
\value{
A Thread object. See https://platform.openai.com/docs/api-reference/threads/object
}
\description{
Retrieve a Thread
}
\examples{
thread_get()
}
